finish airbases routes
