+++ frontend/server/src/main.rs
@@ -1,7 +1,7 @@
 use std::{collections::HashMap, sync::Arc};
 
 use anyhow::Result;
-use axum::{Router, routing::get};
+use axum::Router;
 use enum_iterator::all;
 use once_cell::sync::Lazy;
 use tracing::{error, info};
@@ -35,9 +35,7 @@ fn load_airbases() -> HashMap<Theatre, Airfields> {
 async fn main() -> Result<()> {
     tracing_subscriber::fmt::init();
 
-    let app = Router::new()
-        .route("/api/airbases", get(routes::show_airbases))
-        .route("/api/airbases/{theatre}/{airbase}", get(routes::airbases));
+    let app = Router::new().nest("/api/airbases", routes::airbases::routes());
 
     info!("Starting web server on port 3000");
     let listener = tokio::net::TcpListener::bind("0.0.0.0:3000").await.unwrap();
